Update test to run against new hook
