+++ plugins/woocommerce/client/blocks/assets/js/base/hooks/use-update-preferred-autocomplete-provider.ts
@@ -54,6 +54,7 @@ export function useUpdatePreferredAutocompleteProvider(
 				addressType === 'shipping'
 					? 'shippingAddress'
 					: 'billingAddress';
+
 			return {
 				country: (
 					select(
